requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
pillow==10.3.0
pypng==0.20220715.0
python-resize-image==1.1.20
qrcode==7.4.2
requests==2.32.2
typing_extensions==4.12.0
urllib3==2.2.1
